Fix incorrect player position rounding.
(int) casting rounds towards zero, but float positions need to be
rounded down unconditionally (towards negative infinity) to produce the
correct position.

The problem was that for negative player coordinates the player's block
position would be off by one.

Constructing a new BlockPos() takes care of the correct rounding for us
and makes the code more readable, as that can be used for comparisons.