Fix transparent needles perforating the compass.
Main issue was the needles being rendered before the compass, so the
opaque compass was never drawn where the needles resided and as the
needles are semi-transparent, they left "holes" in the compass through
which you could see the blocks behind it.

Fixed some other minor issues with the rendering as well.