Enable light-level feature and support item tags (#425)
* Add POWER BlockStateProperty to lit blocks

The first step for dynamic lights with variying light level requires the
blockstate to actual hold the light level value. This commit extends the
lit blocks accordingly.

* Allow specification of light level in config files

With this commit each item entry in the dynamiclights_*.cfg can now
specify the light level it emits. Similiar to the pseudo "nameId"
property there is now a "lightLevel" property with type "short".

The defaults with this commit are currently 10 for torches and 15 for
glowstone.

* Allow usage of item tags to make items emit light

With this commit all items having the dynamiclights:self,
dynamiclights:dropped or the dynamiclights:light_[1-15],
dynamiclights_dropped[1-15] item tag now emit light in their
corresponding situations.

* Added item tag for waterproofing of items

Giving the "dynamiclights:not_waterproof" tag to any item that emits
dynamic lights will prevent it to work when submerged in water.